Refactor activities insights helpers
