+++ __tests__/activity-insights-model.test.js
@@ -33,12 +33,7 @@ jest.mock('../public/js/taxonomy/taxonomy-selectors.js', () => ({
     })
 }));
 
-import {
-    buildInsightsModel,
-    buildTrendModel,
-    detectActivityDataIssues,
-    getDefaultTrendDateRange
-} from '../public/js/activities/insights-model.js';
+import { buildInsightsModel } from '../public/js/activities/insights-model.js';
 import { calculateEndDateTime, timeToDateTime } from '../public/js/utils.js';
 
 function isoAt(time) {
@@ -394,152 +389,5 @@ describe('activity insights model', () => {
                 })
             ])
         );
-    });
-
-    test('detectActivityDataIssues detects overlap, invalid-range, and duplicate-auto', () => {
-        const issues = detectActivityDataIssues([
-            activity({
-                id: 'activity-1',
-                startDateTime: isoAt('09:00'),
-                endDateTime: isoAt('10:00'),
-                source: 'task',
-                sourceTaskId: 'task-1'
-            }),
-            activity({
-                id: 'activity-2',
-                startDateTime: isoAt('09:30'),
-                endDateTime: isoAt('10:30'),
-                source: 'manual',
-                sourceTaskId: null
-            }),
-            activity({
-                id: 'activity-3',
-                startDateTime: isoAt('11:00'),
-                endDateTime: isoAt('10:45')
-            }),
-            activity({
-                id: 'activity-4',
-                startDateTime: isoAt('12:00'),
-                endDateTime: isoAt('12:30'),
-                source: 'task',
-                sourceTaskId: 'task-1'
-            })
-        ]);
-
-        expect(issues).toEqual(
-            expect.arrayContaining([
-                expect.objectContaining({
-                    type: 'overlap',
-                    activityId: 'activity-2',
-                    overlappingActivityId: 'activity-1'
-                }),
-                expect.objectContaining({ type: 'invalid-range', activityId: 'activity-3' }),
-                expect.objectContaining({
-                    type: 'duplicate-auto',
-                    activityId: 'activity-4',
-                    sourceTaskId: 'task-1'
-                })
-            ])
-        );
-    });
-
-    test('buildTrendModel builds filtered daily buckets and parent category totals', () => {
-        const model = buildTrendModel({
-            activities: [
-                activity({
-                    id: 'old-work',
-                    startDateTime: isoOn('2026-05-06', '09:00'),
-                    endDateTime: isoOn('2026-05-06', '09:30'),
-                    duration: 30
-                }),
-                activity({
-                    id: 'today-work',
-                    startDateTime: isoAt('09:00'),
-                    endDateTime: isoAt('10:00'),
-                    duration: 60
-                }),
-                activity({
-                    id: 'outside-range',
-                    startDateTime: isoOn('2026-04-23', '09:00'),
-                    endDateTime: isoOn('2026-04-23', '10:00'),
-                    duration: 60
-                })
-            ],
-            now: new Date(isoAt('12:00'))
-        });
-
-        expect(model.dailyHours).toHaveLength(14);
-        expect(model.dailyHours.at(-1)).toEqual({
-            date: '2026-05-07',
-            minutes: 60,
-            categorySegments: [
-                {
-                    key: 'work',
-                    label: 'Work',
-                    color: '#0f172a',
-                    minutes: 60
-                }
-            ]
-        });
-        expect(model.categoryTotals[0]).toEqual({
-            key: 'work',
-            label: 'Work',
-            color: '#0f172a',
-            minutes: 90
-        });
-    });
-
-    test('buildTrendModel splits midnight-crossing activities across local day buckets', () => {
-        const startDateTime = isoOn('2026-05-06', '23:30');
-        const endDateTime = addMinutes(startDateTime, 60);
-
-        const model = buildTrendModel({
-            activities: [
-                activity({
-                    id: 'crossing-activity',
-                    startDateTime,
-                    endDateTime,
-                    duration: 60
-                })
-            ],
-            dateRange: { startDate: '2026-05-06', endDate: '2026-05-07' },
-            now: new Date(isoAt('12:00'))
-        });
-
-        expect(model.dailyHours).toEqual([
-            expect.objectContaining({
-                date: '2026-05-06',
-                minutes: 30,
-                categorySegments: [
-                    expect.objectContaining({
-                        key: 'work',
-                        minutes: 30
-                    })
-                ]
-            }),
-            expect.objectContaining({
-                date: '2026-05-07',
-                minutes: 30,
-                categorySegments: [
-                    expect.objectContaining({
-                        key: 'work',
-                        minutes: 30
-                    })
-                ]
-            })
-        ]);
-        expect(model.categoryTotals[0]).toEqual(
-            expect.objectContaining({
-                key: 'work',
-                minutes: 60
-            })
-        );
-    });
-
-    test('getDefaultTrendDateRange returns the last 14 local days ending today', () => {
-        expect(getDefaultTrendDateRange(new Date(isoAt('12:00')))).toEqual({
-            startDate: '2026-04-24',
-            endDate: '2026-05-07'
-        });
     });
 });
